feat(/update): integrated ORM

+++ backend/src/database/migrations/02-create-packs.js
@@ -34,7 +34,7 @@ module.exports = {
       },
       // QUANTIDADE DO PRODUTO COMPONENTE NO PACK
       qty: {
-        type: Sequelize.DECIMAL(9,2),
+        type: Sequelize.BIGINT,
         allowNull: false,
       },
     });
